+++ src/main/java/org/apiphany/security/client/http/AbstractAuthorizationHttpExchangeClient.java
@@ -1,10 +1,11 @@
-package org.apiphany.security;
+package org.apiphany.security.client.http;
 
 import org.apiphany.ApiRequest;
 import org.apiphany.client.ExchangeClient;
 import org.apiphany.header.Headers;
 import org.apiphany.http.HttpHeader;
 import org.apiphany.lang.ScopedResource;
+import org.apiphany.security.AuthorizationHeaderProvider;
 
 /**
  * Abstract authenticated HTTP exchange client that adds an Authorization header to each request.
@@ -35,7 +36,7 @@ public abstract class AbstractAuthorizationHttpExchangeClient extends AbstractAu
 	 * @see #authenticate(ApiRequest)
 	 */
 	@Override
-	protected final <T> void authenticate(final ApiRequest<T> apiRequest) {
+	public final <T> void authenticate(final ApiRequest<T> apiRequest) {
 		String headerValue = getAuthorizationHeader();
 		Headers.addTo(apiRequest.getHeaders(), HttpHeader.AUTHORIZATION, headerValue);
 	}